% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constrained.R
\name{print.counterfact_parallel}
\alias{print.counterfact_parallel}
\title{Print method for parallel lives}
\usage{
\method{print}{counterfact_parallel}(x, ...)
}
\arguments{
\item{x}{A \code{counterfact_parallel} object.}

\item{...}{Ignored.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print method for parallel lives
}
