% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.counterfactme}
\alias{print.counterfactme}
\title{Print a counterfactual life}
\usage{
\method{print}{counterfactme}(x, ...)
}
\arguments{
\item{x}{A \code{counterfactme} object.}

\item{...}{Ignored.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Pretty-prints the result of \code{\link{counterfact_me}} as a readable
life summary.
}
