% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplers.R
\name{sample_occupation}
\alias{sample_occupation}
\title{Sample a random Norwegian occupation}
\usage{
sample_occupation(n = 1)
}
\arguments{
\item{n}{Number of occupations to draw. Default 1.}
}
\value{
A character vector of occupation names.
}
\description{
Draws one occupation from the 7,020 registered Norwegian occupation codes
(yrkeskoder). All occupations are equally likely by default.
}
\examples{
sample_occupation()
sample_occupation(5)
}
